Complete management of collections

+++ components/view/dialog/delete-collection-alert/delete-collection-alert-dialog-content.tsx
@@ -1,11 +1,12 @@
 "use client";
 
 import { AlertDialogAction, AlertDialogCancel, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle } from "@/components/ui/alert-dialog";
-import { deleteCollection } from "@/hooks/api-endpoints-client";
+import { fetchDeleteCollection } from "@/hooks/api-endpoints-client";
 import { toast } from "@/hooks/use-toast";
 import { useModalStore } from "@/providers/modal";
 import { Trash, X } from "lucide-react";
 import React from "react";
+import { useCollectionContext } from "../../content-collections/collections-context";
 import { DeleteCollectionAlertModalParams } from "./delete-collection-alert-dialog";
 
 export default function DeleteContentAlertDialogContent() {
@@ -14,19 +15,24 @@ export default function DeleteContentAlertDialogContent() {
     );
     const modalParams = useModalStore<DeleteCollectionAlertModalParams | undefined>(state => state.modalParams);
     const [isLoading, startTransition] = React.useTransition();
+    const { states: collections, deleteCollection, addCollection } = useCollectionContext();
 
     const onConfirm = () => {
         if (!modalParams?.id) return;
-        
         startTransition(async () => {
+            const deletedCollection = collections.find(c => c.collection.id === modalParams.id);
+            if (!deletedCollection) return;
+
+            deleteCollection(modalParams?.id);
             try {
-                await deleteCollection(modalParams?.id);
+                await fetchDeleteCollection(modalParams?.id);
                 closeModal();
                 toast({
                     title: "Collection deleted successfully",
                     description: `The collection ${modalParams?.name} has been deleted.`,
                 });
             } catch (error: any) {
+                addCollection(deletedCollection.collection);
                 toast({
                     title: "Failed to delete content",
                     description: error.message ?? "Please try again later.",
